IA/requirements.txt
mysql
mysql-connector-python
python-dotenv